chores: change some folders names and delete some deprecated content
