+++ src/components/animation/product_animation.tsx
